\name{babar-package}
\alias{babar-package}
\alias{babar}
\docType{package}
\title{Bayesian Bacterial Growth Curve analysis in R}
\description{This package is designed to use nested sampling (a Bayesian analysis technique) to compare possible models for bacterial growth curves, as well as extracting parameters.}
\details{
\tabular{ll}{Package: \tab babar\cr
Type: \tab Package\cr
Title: \tab Bayesian Bacterial Growth Curve analysis in R\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-13\cr
Author: \tab Lydia Rickett, Matthew Hartley and Nick Pullen\cr
Maintainer: \tab Matthew Hartley <Matthew.Hartley@jic.ac.uk>\cr
License: \tab GPL-2\cr}
}
\author{Lydia Rickett, Matthew Hartley and Nick Pullen}

\keyword{ package }


