\name{babar-package}
\alias{babar-package}
\alias{babar}
\docType{package}
\title{Bayesian Bacterial Growth Curve analsysis in R}
\description{More about what it does (maybe more than one line)}
\details{
\tabular{ll}{Package: \tab babar\cr
Type: \tab Package\cr
Title: \tab Bayesian Bacterial Growth Curve analsysis in R\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-13\cr
Author: \tab Lydia Rickett, Matthew Hartley and Nick Pullen\cr
Maintainer: \tab Matthew Hartley <Matthew.Hartley@jic.ac.uk>\cr
License: \tab What license is it under?\cr}
}
\author{Lydia Rickett, Matthew Hartley and Nick Pullen}

\keyword{ package }


