\name{babar-package}
\alias{babar-package}
\alias{babar}
\docType{package}
\title{Bayesian Bacterial Growth Curve Analysis in R}
\description{Babar is designed to use nested sampling (a Bayesian analysis technique) to compare possible models for bacterial growth curves, as well as extracting parameters. It allows model evidence and parameter likelihood values to be extracted, and also contains helper functions for comparing distributions as well as direct access to the underlying nested sampling code.}
\details{
\tabular{ll}{Package: \tab babar\cr
Type: \tab Package\cr
Title: \tab Bayesian Bacterial Growth Curve Analysis in R\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-13\cr
Author: \tab Lydia Rickett, Matthew Hartley, Richard Morris and Nick Pullen\cr
Maintainer: \tab Matthew Hartley <Matthew.Hartley@jic.ac.uk>\cr
License: \tab GPL-2\cr
Suggests: \tab knitr, R.rsp\cr
VignetteBuilder: \tab knitr, R.rsp\cr}
}
\author{Lydia Rickett, Matthew Hartley, Richard Morris and Nick Pullen}

\keyword{ package }


